--- src/autoeda/eda.py
@@ -1,155 +1,0 @@
-"""
-Functions to support automated exploratory data analysis (EDA)
-"""
-
-import pandas as pd
-import numpy as np
-import matplotlib.pyplot as plt
-import seaborn as sns
-
-
-def get_unary(df, threshold=0.75, dropna=False):
-    """
-    Returns a list of columns with a single value with a relative
-    frequency percentage over a threshold (75% as default).
-    In case user want only unary columns then threshold should be set to 1.
-
-    Parameters:
-    df: pandas dataframe to analyze.
-    threshold: If the relative frequency of the value with the highest counting
-    is over this threshold then column is selected. Default is 0.75.
-    dropna: If considering or not null values as a value in frequencies.
-    Default is False.
-
-    Returns:
-    unary_cols: list of unary columns.
-    """
-    pass
-
-
-def get_summary_df(df):
-    """
-    Returns a summary dataframe with the main statistics, datatypes,
-    counts/missing values for numerical and categorical columns.
-
-    Parameters:
-    df: dataframe to summarize
-
-    Returns:
-    summary_df: dataframe with the summary statistics
-    """
-    pass
-
-
-def check_na_outliers(
-    df,
-    outlier_method="auto",
-    na_threshold=0.1,
-    outlier_threshold=0.05,
-    return_risk=True,
-    return_suggestions=True,
-):
-    """
-    Diagnose missing values and outliers in a DataFrame.
-
-    This function summarizes the amount of missing values
-    and potential outliers for each column in the input dataset.
-    It is intended for exploratory data analysis and provides
-    interpretable diagnostics for future data preprocessing.
-
-    Parameters
-    ----------
-    df : pandas.DataFrame
-        Input dataset for exploratory data analysis.
-    outlier_method : {"auto", "iqr", "zscore", "mad"}, default="auto"
-        Method used to detect outliers in numeric columns. If "auto",
-        the most appropriate method is selected,
-        based on the distribution of each column.
-    na_threshold : float, default=0.1
-        Proportion threshold above which missing values are flagged
-        as high risk. Further investigation or imputation may be needed
-        if there are lots of missing values.
-    outlier_threshold : float, default=0.05
-        Proportion threshold above which outliers are flagged
-        as high risk.
-    return_risk : bool, default=True
-        Whether to compute and return qualitative risk levels
-        (e.g., "low", "medium", "high") based on the specified
-        thresholds.
-    return_suggestions : bool, default=True
-        Whether to include suggested actions for handling missing
-        values and outliers (e.g., imputation or transformation).
-
-    Returns
-    -------
-    pandas.DataFrame
-        A summary table with one row per column, including
-        missing value statistics, outlier statistics (for numeric
-        columns), and optional risk levels and suggested actions.
-    """
-    pass
-
-
-def plot_correlation_heatmap(df, target=None, method="pearson", figsize=(10, 8)):
-    """
-    Generate a correlation heatmap for numeric features in a dataset.
-
-    This function computes pairwise correlations between numeric columns
-    and visualizes them as a heatmap. Optionally, a target column can be
-    highlighted to show correlations specifically with the target.
-
-    Parameters
-    ----------
-    df : pandas.DataFrame
-        Input dataframe containing numeric and non-numeric features.
-
-    target : str, optional
-        Name of the target column. If provided, correlations with the target
-        may be emphasized or displayed separately.
-
-    method : {"pearson", "spearman", "kendall"}, default="pearson"
-        Correlation method to use.
-
-    figsize : tuple, default=(10, 8)
-        Size of the heatmap figure.
-
-    Returns
-    -------
-    matplotlib.axes.Axes
-        The Axes object containing the heatmap visualization.
-    """
-    pass
-
-
-def plot_histograms_by_target(df, target, features=None, bins=30, figsize=(12, 8)):
-    """
-    Plot histograms of features grouped by a target variable.
-
-    This function visualizes the distribution of selected features
-    conditioned on the target variable. This is useful for understanding
-    class separation and feature behavior.
-
-    Parameters
-    ----------
-    df : pandas.DataFrame
-        Input dataframe containing features and target.
-
-    target : str
-        Name of the target column.
-
-    features : list of str, optional
-        List of feature column names to plot. If None, all numeric
-        columns except the target will be used.
-
-    bins : int, default=30
-        Number of bins for the histograms.
-
-    figsize : tuple, default=(12, 8)
-        Size of the figure.
-
-    Returns
-    -------
-    matplotlib.figure.Figure
-        Figure object containing the histograms.
-    """
-    pass
